Arduino SonicSensor_Version.1
The first part of the arduino alarm system using sonic sensors is complete. When the alarm isn't going off a green led will stay on. But when the sonic sensor picks something up and is within a certain range, the red leds will blink while the buzzer makes sound and this will repeat itself untill whatever is in front of the sensor is gone.

The second version will have 2 more components which is a button and a liquid crystal screen.